16-bit bone selectors, so a rig may have more than 256 joints
FModelVertex::boneselector goes from uint8_t[4] to uint16_t[4], with a new
VFmt_UShort4_UInt vertex format to carry it. The IQM loader reads USHORT bone
index arrays, which is what our exporter writes for a rig the one-byte IQM
convention cannot address -- the Slayer has 924 joints.

MD3_MAX_SURFACES also goes 32 -> 64.

AND THE PART THAT MATTERS MORE THAN THE FEATURE.

The three backends each keep a POSITIONAL table indexed by that enum, in three
separate files with nothing linking them. This landed with GL and GLES updated
and VULKAN MISSED -- which is the only backend VR runs on. There was no compile
error and no log line: the backend indexed one past the end of its table and
handed the GPU a garbage vertex format. It surfaced as

    Execution could not continue.
    Could not submit command buffer: device lost

which names neither the format, nor the table, nor the file, and looks for all
the world like a broken driver or a bad model.

So the enum now ends in VFmt_COUNT and all three tables static_assert against
it. Add a format and forget a backend and the build fails, naming the file.
That guard is the actual deliverable here; the 16-bit selectors are 27 lines.

Verified in the headset: E1M1 loads with ModelSwapper's models, no device loss.

+++ src/common/rendering/i_modelvertexbuffer.h
@@ -26,7 +26,7 @@ struct FModelVertex
 	unsigned packedNormal;	// normal vector as GL_INT_2_10_10_10_REV.
 	float lu, lv;	// lightmap texture coordinates
 	float lindex;	// lightmap texture index
-	uint8_t boneselector[4];
+	uint16_t boneselector[4];	// [XR] 16-bit so a rig may have more than 256 joints (VFmt_UShort4_UInt)
 	uint8_t boneweight[4];
 
 	void Set(float xx, float yy, float zz, float uu, float vv)
